added functionality to join users and results
