+++ src/Components/Pages/Admin/Accordion/index.js
@@ -1,12 +1,50 @@
 import Accordion from 'react-bootstrap/Accordion';
 import Table from 'react-bootstrap/Table';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import AdminModal from '../Modal';
+import fetchApi from '../../../../Hooks/useFetch';
+import useJoin from '../../../../Hooks/useJoin';
 
 const TableAccordion = (props) => {
+    // Set useState for results, show and questions
+    const [results, setResults] = useState(null);
+    const [questions, setQuestions] = useState(null);
     const [show, setShow] = useState(false);
+    const [adminModal, setModal] = useState(null);
     const handleShow = () => setShow(true);
     const handleClose = () => setShow(false);
+
+    // On mount async and await fetching results and questions from API
+    useEffect(async () => {
+        await getResults();
+        await getQuestions();
+    }, []);
+    // Make the functionality to fetch results and questions from the API
+    const getResults = async () => {
+        let response = await fetchApi(`result`);
+        if (response.success) {
+            return setResults(response.data);
+        }
+    };
+    const getQuestions = async () => {
+        let response = await fetchApi(`question`);
+        if (response.success) {
+            return setQuestions(response.data);
+        }
+    };
+    useEffect(() => {
+        questions && results
+            ? setModal(
+                  <AdminModal
+                      show={show}
+                      setShow={setShow}
+                      onHide={handleClose}
+                      results={results}
+                      questions={questions}
+                  />,
+              )
+            : '';
+    }, [questions, results]);
     return (
         <Accordion>
             <Accordion.Item eventKey="0">
@@ -31,7 +69,7 @@ const TableAccordion = (props) => {
                                 <td>{props.user.score}</td>
                                 <td>
                                     <button onClick={handleShow}>See answers</button>
-                                    <AdminModal show={show} onHide={handleClose} />
+                                    {adminModal}
                                 </td>
                             </tr>
                         </tbody>
